requirements.txt
flask==2.0.1
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.5.0
scikit-learn==1.0.2
transformers==4.21.0
torch==1.9.0
werkzeug==2.0.1